Time: 4 ms (30.86%) | Memory: 71.7 MB (90.40%) - LeetSync

+++ 1445-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold/number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold.java
@@ -5,7 +5,7 @@ class Solution {
         for (int j = 0; j < arr.length; j++) {
             sum += arr[j];
             if (j >= k - 1) {
-                double avg = sum / k;
+                int avg = sum / k;
                 if (avg >= threshold)
                     res++;
                 sum -= arr[i++];
